requirements.txt
datasets>=2.15.0
Pillow>=10.0.0
requests>=2.31.0
anthropic>=0.8.0
python-dotenv>=1.0.0  